refactor: Subscribe feature
